+++ R/nice_lavaanPlot.R
@@ -389,6 +389,8 @@ nice_lavaanPlot <- function(
     graph_options$labelloc <- if (has_title) "t" else "b"
     # Center the label horizontally
     graph_options$labeljust <- "c"
+    # Center the entire graph drawing on the page
+    graph_options$center <- "true"
   }
 
   lavaanPlot::lavaanPlot(